refac: music-controls now stores all relevant player info inside of track field

+++ src-tauri/crates/app/src/discord.rs
@@ -19,7 +19,7 @@ pub struct PayloadData {
     pub album_cover: Option<String>,
     pub show_timestamps: bool,
     pub progress: f64,
-    pub duration: f32,
+    pub duration: f64,
 }
 
 impl DiscordState {
